Buff Encumberance on Scout Suits
Based on my math, the Survivor scout suit is still 3 times as encumbering as the AEP suit and the Survivor Armored Scout is as encumbering as the ANBC suit, and even more so when full of items. I think this makes more sense in vanilla balance, and allows the suits to be melee viable.